backend/requirements.txt
Flask==1.0.2
flask-sqlalchemy
Flask-Migrate
pymysql
SQLAlchemy-Utils
requests
PyJWT
pylint
flask-redis
psycopg2
gunicorn
psycopg2-binary
python-dotenv